Script changes and non-mac lazygit config

+++ safe-copy.nu
@@ -14,6 +14,14 @@ export def safe_copy [source: string, destination: string] {
   if (is_file $destination) {
     let old_file = $"($destination).old"
 
+    if (is_file $old_file) {
+      print $"File '($old_file) exists. Think carefully before answering: Overwrite '($old_file)' with the contents of the current file?"
+      let answer = input --default "n" "[y/N]" | str downcase
+      if ($answer != 'y') {
+        return 1
+      }
+    }
+
     print $"Destination file '($destination)' exists. Renaming to '($old_file)'."
     mv $destination $old_file
     print "Successfully renamed existing file."
